refactor (Amount) : Getter를 지양하기 위해 동등성 비교를 하도록 리팩토링 한다
- getAmount()를 사용하여 값을 비교하는 대신 객체와 객체를 비교
